requirements.txt
Django==1.7.6
django-filter==0.9.2
djangorestframework==3.1.0
MySQL-python==1.2.5
